+++ CHANGELOG.md
@@ -10,6 +10,7 @@ List all changes after the last release here (newer on top). Each change on a se
 
 ### Changed
 
+- Importer: add context object while initializing a importer class
 - Core: use UUID in basket IDs to prevent possible duplicates
 - Core: save basket shipping and billing address as dictionary when id is not available
 - Front: remove the custom _load() implementation from the basket as it is the same as the core
@@ -24,9 +25,12 @@ List all changes after the last release here (newer on top). Each change on a se
 - Core: add shops and supplier to ServiceProvider and Service models
 - Front: add feature for checkout phases to spawn extra phases
 - Add custom get_ip method and use it everywhere
+- Importer: add permissions for all the diffrent types of importers
+- Importer: add context class to data importer
 
 ### Fixed
 
+- Fixed the way the permissions identifier are split in admin
 - Fixed issue that was importing User model directly
 - Core: changed `del` basket command handler to not try to parse the basket line into an integer